requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain
langchain-nomic
langchain-pinecone
langchain-openai
jq
unstructured
tiktoken
pinecone
pandas
numpy
pillow
google-genai
beautifulsoup4